+++ app/Services/UserService.php
@@ -10,23 +10,33 @@ use Illuminate\Support\Facades\Hash;
 
 class UserService{
 
-    public function isAdmin(int $id){
-        $user = User::find($id);
-        return $user->admin;
-    }
-
-    public function showUsers(){
+    public function showUsers()
+    {
         return User::IdAsc()->paginate(10);
     }
-    public function getUsers(){
+
+    public function getUsers()
+    {
         return User::all();
     }
-    public function getUserById(int $id){
-        return User::find($id);
+
+    public function getUserById(int $id)
+    {
+        $user = User::find($id);
+        if(!$user){
+            abort(404);
+        }
+        return $user;
     }
 
-    public function addUser(AddUserRequest $request){
+    public function addUser(AddUserRequest $request)
+    {
         $user = new User();
+        $this->saveUser($user, $request);
+    }
+
+    public function saveUser(User $user, $request)
+    {
         $user->name = $request->name;
         $user->email = $request->email;
         $user->password = Hash::make($request->password);
@@ -37,16 +47,18 @@ class UserService{
     public function updateUser($request, int $id)
     {
         $user = User::find($id);
-        $user->name = $request->name;
-        $user->email = $request->email;
-        $user->password = Hash::make($request->password);
-        $user->admin = boolval($request->admin);
-        $user->save();
+        if(!$user){
+            abort(404);
+        }
+        $this->saveUser($user, $request);
     }
 
-    public function deleteUser(int $id){
+    public function deleteUser(int $id)
+    {
         $user = User::find($id);
+        if(!$user){
+            abort(404);
+        }
         $user->delete();
     }
-
 }
